+++ robo_gui/src/presentation/presenters/status_presenter/status_presenter.h
@@ -18,9 +18,9 @@ namespace presentation
         Q_PROPERTY(bool isCharging READ isCharging NOTIFY isChargingChanged)
         Q_PROPERTY(int gamepadBatteryLevel READ gamepadBatteryLevel NOTIFY gamepadBatteryLevelChanged)
         Q_PROPERTY(int robotBatteryLevel READ robotBatteryLevel NOTIFY robotBatteryLevelChanged)
-        Q_PROPERTY(int gunPositionH READ gunPositionH NOTIFY gunPositionHChanged)
-        Q_PROPERTY(int gunPositionV READ gunPositionV NOTIFY gunPositionVChanged)
-        Q_PROPERTY(int cameraPositionV READ cameraPositionV NOTIFY cameraPositionVChanged)
+        Q_PROPERTY(qreal gunPositionH READ gunPositionH NOTIFY gunPositionHChanged)
+        Q_PROPERTY(qreal gunPositionV READ gunPositionV NOTIFY gunPositionVChanged)
+        Q_PROPERTY(qreal cameraPositionV READ cameraPositionV NOTIFY cameraPositionVChanged)
         Q_PROPERTY(qreal yaw READ yaw NOTIFY yawChanged)
         Q_PROPERTY(qreal pitch READ pitch NOTIFY pitchChanged)
         Q_PROPERTY(qreal roll READ roll NOTIFY rollChanged)
@@ -35,9 +35,9 @@ namespace presentation
         bool isCharging() const;
         int gamepadBatteryLevel() const;
         int robotBatteryLevel() const;
-        int gunPositionH() const;
-        int gunPositionV() const;
-        int cameraPositionV() const;
+        qreal gunPositionH() const;
+        qreal gunPositionV() const;
+        qreal cameraPositionV() const;
         qreal yaw() const;
         qreal pitch() const;
         qreal roll() const;
@@ -51,9 +51,9 @@ namespace presentation
         void isChargingChanged(bool charging);
         void gamepadBatteryLevelChanged(int level);
         void robotBatteryLevelChanged(int level);
-        void gunPositionHChanged(int position);
-        void gunPositionVChanged(int position);
-        void cameraPositionVChanged(int position);
+        void gunPositionHChanged(qreal position);
+        void gunPositionVChanged(qreal position);
+        void cameraPositionVChanged(qreal position);
         void yawChanged(qreal yaw);
         void pitchChanged(qreal pitch);
         void rollChanged(qreal roll);
